requirements.txt
# Core Flask
Flask==3.0.3
Werkzeug==3.0.3
Jinja2==3.1.2
itsdangerous==2.2.0
gunicorn==20.1.0

# Banco de dados
psycopg2-binary==2.9.10

# AWS S3
boto3==1.38.13
botocore==1.38.13
s3transfer==0.12.0
jmespath==1.0.1

# Variáveis de ambiente
python-dotenv==1.1.0

# Utilitários
python-dateutil==2.8.2
requests==2.28.1
urllib3==1.26.12